Try to identify if day0.exs is run as a script, instead of a separate run.ex file
Elixir doesn't have a "main method" or "check if this is the main file"
feature, and the recommended approach is to use mix to build an escript.
But adding mix to AoC seems overkill, and it's nice to be able to run
a day program directly, but not try to read stdin when loading the file
in an iex session.  Erlang's escript.script_name() appears to do
something like C's arvg[0], and returns "./day0.exs" when a script is
invoked like `./day0.exs` or `elixir day0/day0.exs` which can then be
compared with __ENV__.file.  Unfortunately, Erlang docs say[1]
"If the function is invoked outside the context of an escript, the
behavior is undefined."  Observed behavior is that
`elixir -r runner.exs day0/day0.exs` causes script_name to return "-r"
and likewise `elixir -S day0.exs` returns "-S".  If invoking a script
that way, use `elixir -r day0.exs -e 'Day0.main'`.  (I considered a
separate `Main` module, but that would result in warnings if multiple
days were built into a single program.)

[1] https://www.erlang.org/doc/man/escript.html#script_name-0